requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
numpy
Pillow
bcrypt==4.1.2
google-generativeai==0.3.2
scikit-learn
scipy
cmake==3.27.7
opencv-python==4.9.0.80
